fix: harden integration response redaction
