+++ src/web-console/stores/ConsoleStoreValidation.ts
@@ -100,6 +100,27 @@ export function assertNullableDisplayString(value: string | null, name: string, 
   assertDisplayString(value, name, maxLength);
 }
 
+export function isWellFormedUnicode(value: string): boolean {
+  for (let index = 0; index < value.length; index += 1) {
+    const codeUnit = value.charCodeAt(index);
+    if (codeUnit >= 0xd800 && codeUnit <= 0xdbff) {
+      if (index + 1 >= value.length) return false;
+      const next = value.charCodeAt(index + 1);
+      if (next < 0xdc00 || next > 0xdfff) return false;
+      index += 1;
+    } else if (codeUnit >= 0xdc00 && codeUnit <= 0xdfff) {
+      return false;
+    }
+  }
+  return true;
+}
+
+export function assertWellFormedUnicode(value: string, name: string): void {
+  if (!isWellFormedUnicode(value)) {
+    throw new ConsoleStoreValidationError(`${name} must contain well-formed Unicode`);
+  }
+}
+
 /**
  * Membership test for a readonly literal tuple against a wider string.
  * `tuple.includes(value)` won't type-check when value is a wider `string`
